ci: bump cargo-audit to v0.18.1 (#123)
The new cargo-audit depends on tame-index, instead of crates-index.
The tame-index package uses gix to fetch/read a git registry index.
So, deleted libssh2 and libgit2 because no longer necessary
to build either of them for cargo-audit.

n.b., openssl and libz-sys are likely to have some usages in the future,
so leaving them as they are.